Move domain enrichment worker within a separate directory
