+++ worker/domain_enrichment/domain_enrichment_worker.ts
@@ -11,7 +11,7 @@ import {
 import { prisma } from '@/lib/prisma';
 import { logger } from '@/lib/logging/logger';
 import { ClaimedDomainJob } from '@/dal/domain_enrichment_jobs/domain_enrichment_jobs.types';
-import { DomainEnrichmentJobResult } from '@/worker/domain_enrichment_worker_types';
+import { DomainEnrichmentJobResult } from '@/worker/domain_enrichment/domain_enrichment_worker_types';
 import { sleep } from '@/lib/timeouts';
 import { msSince } from '@/lib/time';
 import { normalizeError } from '@/lib/errors/utils';
